Grok Disinformation Checker 2/requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
httpx
newspaper3k
lxml_html_clean